requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
langchain>=0.2.11
langgraph>=0.1.14
sqlalchemy>=2.0.29
aiosqlite>=0.19.0
pytest>=8.1.0
ruff>=0.4.0
mypy>=1.8.0
PyYAML>=6.0
